Commands: APICommand made an interface, and renamed to BitfinexCommand
