+++ src/main/java/com/github/jnidzwetzki/bitfinex/v2/command/CalculateCommand.java
@@ -15,29 +15,28 @@
  *    limitations under the License. 
  *    
  *******************************************************************************/
-package com.github.jnidzwetzki.bitfinex.v2.commands;
-
-import org.json.JSONObject;
+package com.github.jnidzwetzki.bitfinex.v2.command;
 
 import com.github.jnidzwetzki.bitfinex.v2.BitfinexApiBroker;
-import com.github.jnidzwetzki.bitfinex.v2.symbol.BitfinexCandlestickSymbol;
 
-public class SubscribeCandlesCommand extends AbstractAPICommand {
+public class CalculateCommand implements BitfinexCommand {
+	
+	/**
+	 * The symbol
+	 */
+	private String symbol;
 
-	private final BitfinexCandlestickSymbol symbol;
-
-	public SubscribeCandlesCommand(final BitfinexCandlestickSymbol symbol) {
+	public CalculateCommand(final String symbol) {
 		this.symbol = symbol;
 	}
 
 	@Override
 	public String getCommand(final BitfinexApiBroker bitfinexApiBroker) {
-		final JSONObject subscribeJson = new JSONObject();
-		subscribeJson.put("event", "subscribe");
-		subscribeJson.put("channel", "candles");
-		subscribeJson.put("key", symbol.toBifinexCandlestickString());
-				
-		return subscribeJson.toString();
+		final StringBuilder sb = new StringBuilder();
+		sb.append("[0,\"calc\",null,[[\"");
+		sb.append(symbol);
+		sb.append("\"]]]");
+		
+		return sb.toString();
 	}
-
 }
