+++ src/main/java/com/github/jnidzwetzki/bitfinex/v2/command/SubscribeTickerCommand.java
@@ -15,39 +15,28 @@
  *    limitations under the License. 
  *    
  *******************************************************************************/
-package com.github.jnidzwetzki.bitfinex.v2.commands;
+package com.github.jnidzwetzki.bitfinex.v2.command;
 
 import org.json.JSONObject;
 
 import com.github.jnidzwetzki.bitfinex.v2.BitfinexApiBroker;
-import com.github.jnidzwetzki.bitfinex.v2.exception.CommandException;
-import com.github.jnidzwetzki.bitfinex.v2.symbol.BitfinexOrderBookSymbol;
+import com.github.jnidzwetzki.bitfinex.v2.symbol.BitfinexTickerSymbol;
 
-public class SubscribeOrderbookCommand extends AbstractAPICommand {
+public class SubscribeTickerCommand implements BitfinexCommand {
 
-	/**
-	 * The orderbook configuration
-	 */
-	private BitfinexOrderBookSymbol symbol;
-	
-	public SubscribeOrderbookCommand(final BitfinexOrderBookSymbol symbol) {
-		this.symbol = symbol;
+	private String currencyPair;
+
+	public SubscribeTickerCommand(final BitfinexTickerSymbol currencyPair) {
+		this.currencyPair = currencyPair.getBitfinexCurrencyPair().toBitfinexString();
 	}
 
 	@Override
-	public String getCommand(final BitfinexApiBroker bitfinexApiBroker) throws CommandException {
+	public String getCommand(final BitfinexApiBroker bitfinexApiBroker) {
 		final JSONObject subscribeJson = new JSONObject();
 		subscribeJson.put("event", "subscribe");
-		subscribeJson.put("channel", "book");
-		subscribeJson.put("symbol", symbol.getCurrencyPair().toBitfinexString());
-		subscribeJson.put("prec", symbol.getPrecision().toString());
-		if (symbol.getFrequency() != null) {
-			subscribeJson.put("freq", symbol.getFrequency().toString());
-		}
-		if (symbol.getPricePoints() != null) {
-			subscribeJson.put("len", Integer.toString(symbol.getPricePoints()));
-		}
+		subscribeJson.put("channel", "ticker");
+		subscribeJson.put("symbol", currencyPair);
+		
 		return subscribeJson.toString();
 	}
-
 }
