+++ src/test/java/com/github/jnidzwetzki/bitfinex/v2/test/CommandsTest.java
@@ -11,20 +11,20 @@ import org.mockito.Mockito;
 import com.github.jnidzwetzki.bitfinex.v2.BitfinexApiBroker;
 import com.github.jnidzwetzki.bitfinex.v2.BitfinexApiBrokerConfig;
 import com.github.jnidzwetzki.bitfinex.v2.BitfinexOrderBuilder;
-import com.github.jnidzwetzki.bitfinex.v2.commands.AbstractAPICommand;
-import com.github.jnidzwetzki.bitfinex.v2.commands.AuthCommand;
-import com.github.jnidzwetzki.bitfinex.v2.commands.CancelOrderCommand;
-import com.github.jnidzwetzki.bitfinex.v2.commands.CancelOrderGroupCommand;
+import com.github.jnidzwetzki.bitfinex.v2.command.BitfinexCommand;
+import com.github.jnidzwetzki.bitfinex.v2.command.AuthCommand;
+import com.github.jnidzwetzki.bitfinex.v2.command.CancelOrderCommand;
+import com.github.jnidzwetzki.bitfinex.v2.command.CancelOrderGroupCommand;
 import com.github.jnidzwetzki.bitfinex.v2.entity.BitfinexNewOrder;
 import com.github.jnidzwetzki.bitfinex.v2.exception.CommandException;
-import com.github.jnidzwetzki.bitfinex.v2.commands.OrderCommand;
-import com.github.jnidzwetzki.bitfinex.v2.commands.PingCommand;
-import com.github.jnidzwetzki.bitfinex.v2.commands.SetConnectionFeaturesCommand;
-import com.github.jnidzwetzki.bitfinex.v2.commands.SubscribeCandlesCommand;
-import com.github.jnidzwetzki.bitfinex.v2.commands.SubscribeOrderbookCommand;
-import com.github.jnidzwetzki.bitfinex.v2.commands.SubscribeTickerCommand;
-import com.github.jnidzwetzki.bitfinex.v2.commands.SubscribeTradesCommand;
-import com.github.jnidzwetzki.bitfinex.v2.commands.UnsubscribeChannelCommand;
+import com.github.jnidzwetzki.bitfinex.v2.command.OrderCommand;
+import com.github.jnidzwetzki.bitfinex.v2.command.PingCommand;
+import com.github.jnidzwetzki.bitfinex.v2.command.SetConnectionFeaturesCommand;
+import com.github.jnidzwetzki.bitfinex.v2.command.SubscribeCandlesCommand;
+import com.github.jnidzwetzki.bitfinex.v2.command.SubscribeOrderbookCommand;
+import com.github.jnidzwetzki.bitfinex.v2.command.SubscribeTickerCommand;
+import com.github.jnidzwetzki.bitfinex.v2.command.SubscribeTradesCommand;
+import com.github.jnidzwetzki.bitfinex.v2.command.UnsubscribeChannelCommand;
 import com.github.jnidzwetzki.bitfinex.v2.entity.BitfinexCurrencyPair;
 import com.github.jnidzwetzki.bitfinex.v2.entity.BitfinexOrderType;
 import com.github.jnidzwetzki.bitfinex.v2.entity.BitfinexCandleTimeFrame;
@@ -55,7 +55,7 @@ public class CommandsTest {
 
 		BitfinexOrderBookSymbol rawOrderbookConfiguration = BitfinexSymbols.rawOrderBook(BitfinexCurrencyPair.of("BAT", "BTC"));
 
-		final List<AbstractAPICommand> commands = Arrays.asList(
+		final List<BitfinexCommand> commands = Arrays.asList(
 				new AuthCommand(AuthCommand.AUTH_NONCE_PRODUCER_TIMESTAMP),
 				new CancelOrderCommand(123),
 				new CancelOrderGroupCommand(1),
@@ -71,7 +71,7 @@ public class CommandsTest {
 
 		final BitfinexApiBroker bitfinexApiBroker = buildMockedBitfinexConnection();
 
-		for(final AbstractAPICommand command : commands) {
+		for(final BitfinexCommand command : commands) {
 			if (command instanceof BitfinexStreamSymbolToChannelIdResolverAware) {
 				((BitfinexStreamSymbolToChannelIdResolverAware) command).setResolver(s -> 12);
 			}
